+++ components/dropdown/dropdown.vue
@@ -8,6 +8,7 @@
     :fallback-placements="fallbackPlacements"
     padding="none"
     role="menu"
+    :append-to="appendTo"
     :modal="modal"
     :max-height="maxHeight"
     :max-width="maxWidth"
@@ -57,6 +58,9 @@ import { LIST_ITEM_NAVIGATION_TYPES } from '../list_item/list_item_constants';
 import {
   DROPDOWN_PADDING_CLASSES,
 } from './dropdown_constants';
+import {
+  POPOVER_APPEND_TO_VALUES,
+} from '@/components/popover/popover_constants';
 import { getUniqueString } from '@/common/utils';
 import { EVENT_KEYNAMES } from '@/common/constants';
 import SrOnlyCloseButtonMixin from '@/common/mixins/sr_only_close_button';
@@ -107,6 +111,7 @@ export default {
 
     /**
      * Vertical padding size around the list element.
+     * @values none, small, large
      */
     padding: {
       type: String,
@@ -128,6 +133,7 @@ export default {
     /**
      * Width configuration for the popover content. When its value is 'anchor',
      * the popover content will have the same width as the anchor.
+     * @values null, anchor
      */
     contentWidth: {
       type: String,
@@ -166,6 +172,7 @@ export default {
      * - "arrow-keys" for items that are navigated with UP/DOWN keys.
      * - "tab" for items that are navigated using the TAB key.
      * - "none" for static items that are not interactive.
+     * @values arrow-keys, tab, none
      */
     navigationType: {
       type: String,
@@ -176,6 +183,12 @@ export default {
     /**
      * If the dropdown does not fit in the direction described by "placement",
      * it will attempt to change it's direction to the "fallbackPlacements".
+     *
+     * @values top, top-start, top-end,
+     * right, right-start, right-end,
+     * left, left-start, left-end,
+     * bottom, bottom-start, bottom-end,
+     * auto, auto-start, auto-end
      * */
     fallbackPlacements: {
       type: Array,
@@ -214,6 +227,20 @@ export default {
     listClass: {
       type: [String, Array, Object],
       default: '',
+    },
+
+    /**
+     * Sets the element to which the popover is going to append to.
+     * 'body' will append to the nearest body (supports shadow DOM).
+     * @values 'body', 'parent', HTMLElement,
+     */
+    appendTo: {
+      type: [HTMLElement, String],
+      default: 'body',
+      validator: appendTo => {
+        return POPOVER_APPEND_TO_VALUES.includes(appendTo) ||
+            (appendTo instanceof HTMLElement);
+      },
     },
   },