+++ recipes/comboboxes/combobox_with_popover/combobox_with_popover.stories.js
@@ -113,6 +113,14 @@ export const argTypesData = {
       },
     },
   },
+  appendTo: {
+    defaultValue: 'body',
+    table: {
+      defaultValue: {
+        summary: 'body',
+      },
+    },
+  },
   padding: {
     defaultValue: 'small',
     control: {