+++ recipes/leftbar/general_row/leftbar_general_row_icon.vue
@@ -1,9 +1,8 @@
 <template>
   <dt-icon
-    v-if="shouldRenderWithIcon"
+    v-if="isIconType"
     :name="getIconName"
-    :class="iconColor"
-    :size="isDialbotType ? '500' : '300'"
+    size="300"
   />
   <div
     v-else-if="isContactCenterType"
@@ -14,7 +13,7 @@
     :class="dialbotClasses"
   >
     <icon-dialpad-glyph
-      class="d-svg--size16"
+      class="d-svg--size18"
       color="white"
     />
   </div>
@@ -24,9 +23,9 @@
 import { DtIcon } from '@/components/icon';
 import IconDialpadGlyph from '@dialpad/dialtone/lib/dist/vue/icons/IconDialpadGlyph';
 import {
-  LEFTBAR_GENERAL_ROW_ICON_MAPPING,
-  LEFTBAR_GENERAL_ROW_TYPES,
-  LEFTBAR_GENERAL_ROW_CONTACT_CENTER_COLORS,
+  LEFTBAR_GENERAL_ROW_ICON_MAPPING as ICON_MAPPING,
+  LEFTBAR_GENERAL_ROW_TYPES as TYPES,
+  LEFTBAR_GENERAL_ROW_CONTACT_CENTER_COLORS as COLORS,
 } from '@/recipes/leftbar/general_row/general_row_constants';
 
 export default {
@@ -45,59 +44,40 @@ export default {
   },
 
   computed: {
-    shouldRenderWithIcon () {
-      return this.type !== LEFTBAR_GENERAL_ROW_TYPES.DIALBOT && this.type !== LEFTBAR_GENERAL_ROW_TYPES.CONTACT_CENTER;
+    isIconType () {
+      return ![TYPES.DIALBOT, TYPES.CONTACT_CENTER].includes(this.type);
+    },
+
+    isContactCenterType () {
+      return this.type === TYPES.CONTACT_CENTER;
+    },
+
+    isDialbotType () {
+      return this.type === TYPES.DIALBOT;
+    },
+
+    getIconName () {
+      return ICON_MAPPING[this.type];
     },
 
     contactCenterIconClasses () {
       return [
-        'd-icon',
-        'leftbar-general-row__contact-center-icon',
-        LEFTBAR_GENERAL_ROW_CONTACT_CENTER_COLORS[this.color],
+        'dt-leftbar-row__icon-cc',
+        COLORS[this.color],
       ];
     },
 
     dialbotClasses () {
       return [
-        'leftbar-general-row__dialbot-icon',
         'd-bar-circle',
+        'd-bgc-purple-400',
+        'd-w24',
+        'd-h24',
+        'd-d-flex',
+        'd-ai-center',
+        'd-jc-center',
       ];
-    },
-
-    iconColor () {
-      return this.isContactCenterType
-        ? LEFTBAR_GENERAL_ROW_CONTACT_CENTER_COLORS[this.color]
-        : 'd-fc-black-700';
-    },
-
-    isContactCenterType () {
-      return this.type === LEFTBAR_GENERAL_ROW_TYPES.CONTACT_CENTER;
-    },
-
-    isDialbotType () {
-      return this.type === LEFTBAR_GENERAL_ROW_TYPES.DIALBOT;
-    },
-
-    getIconName () {
-      return LEFTBAR_GENERAL_ROW_ICON_MAPPING[this.type];
     },
   },
 };
 </script>
-
-<style scoped>
-.leftbar-general-row__contact-center-icon {
-  width: var(--icon-size-200);
-  height: var(--icon-size-200);
-  border-radius: var(--size-200);
-}
-
-.leftbar-general-row__dialbot-icon {
-  background-color: var(--purple-400);
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-</style>